consistent inter-quote breaks for old MIT quotes
I picked the style that had been adopted towards the end of the old file, and
applied that globally.  The ideas was to make it clear where the quotes start on
finish, because - even after removing the double-spacing - there are some with
paragraph breaks or fixed formatting that is important to the quote, but makes
surrounding quotes seem to run on into them.

Also:  OCD.