Convert 4-bits-part3 to Markdown for consistent formatting
All the earlier 4-bits posts were Markdown, but Part 3 was IPython.  I was
thinking to demonstrate some of Jupyter's pop-up help, but it turns out that
this pop-up help does indeed pop up, not going inline of the document. I guess
that's a good thing, but it meant there was no point in using the notebook.

Also the inline code formatting is a bit nicer with the Markdown format.