Turn on TAG_PAGES_ARE_INDEXES
When you view a tag page, there's no summary of the posts unless
TAG_PAGES_ARE_INDEXES is True.

So now all my Tags have an index page with the teaser text, not just the main
blog index.

This is particularly nice for the 4-bit-rules tag where all the titles are very
similar: with the teaser text you can see what each post is actually about.