customise milosophy theme:  full-width container
Use all of the window width please, not just 80% of it.  This millenial habbit
of putting wide margins on web pages is stupid.

If I want a narrow column then either I would make the window narrow. Conversly
on a phone, it's already a narrow column in portrait, and if I want a wide
column then I would turn the phone sideways and expect the whole screen to be
used!